requirements.txt
grpcio-tools
ultralytics